Refactor profile and chat components, so that separation of concerns is tangible
